+++ src/Manopt.jl
@@ -448,6 +448,7 @@ export NeverRestart, RestartOnNonDescent, RestartOnNonSufficientDescent
 #
 # Solvers
 export adaptive_regularization_with_cubics,
+    accepted_keywords,
     adaptive_regularization_with_cubics!,
     alternating_gradient_descent,
     alternating_gradient_descent!,
@@ -481,7 +482,6 @@ export adaptive_regularization_with_cubics,
     gradient_descent!,
     interior_point_Newton,
     interior_point_Newton!,
-    keywords_accepted,
     LevenbergMarquardt,
     LevenbergMarquardt!,
     mesh_adaptive_direct_search,